# vectorize slice::is_sorted