@Kobzol 2024-11-11
FWIW, the two changed benchmarks are most likely spurious, they are oscillating currently. So it's perf. neutral, which is fine.