@therealprof 2024-11-13
It would be nice if we could check impacts not only on x86_64 but also (at least) on aarch64. The original function was already partially vectorized on x86_64 but the new generated code is growing substantially on ARMv7 and to a lesser extent on aarch64 and I do have doubts about this being a universal improvement...