@LaihoE 2024-11-13
I'm completely fine with reverting this if there are concerns with other targets. The improvements on x86 are not that big.

>The original function was already partially vectorized on x86_64

Do you have an example of this? I would expect that optimization to be illegal as the old function used ```all()``` that short circuits on every element.

I must admit that my experience with aarch64 is quite limited but isn't this a fine hot loop?
```
.LBB0_4:
        ldp     q1, q2, [x10, #-32]
        ldur    q3, [x10, #-28]
        ldur    q4, [x10, #-12]
        ldur    q17, [x10, #-76]
        ldur    q18, [x10, #-60]
        ldur    q19, [x10, #-44]
        ldur    q20, [x10, #-124]
        ldur    q21, [x10, #-108]
        cmgt    v2.4s, v2.4s, v4.4s
        cmgt    v1.4s, v1.4s, v3.4s
        ldur    q22, [x10, #-92]
        ldp     q5, q6, [x10, #-128]
        ldp     q3, q4, [x10, #-96]
        uzp1    v1.8h, v1.8h, v2.8h
        ldp     q7, q16, [x10, #-64]
        cmgt    v6.4s, v6.4s, v21.4s
        cmgt    v5.4s, v5.4s, v20.4s
        cmgt    v4.4s, v4.4s, v17.4s
        cmgt    v3.4s, v3.4s, v22.4s
        cmgt    v16.4s, v16.4s, v19.4s
        cmgt    v7.4s, v7.4s, v18.4s
        uzp1    v3.8h, v3.8h, v4.8h
        uzp1    v4.8h, v5.8h, v6.8h
        uzp1    v2.8h, v7.8h, v16.8h
        uzp1    v1.16b, v2.16b, v1.16b
        uzp1    v2.16b, v4.16b, v3.16b
        and     v1.16b, v1.16b, v0.16b
        and     v2.16b, v2.16b, v0.16b
        ext     v3.16b, v1.16b, v1.16b, #8
        ext     v4.16b, v2.16b, v2.16b, #8
        zip1    v1.16b, v1.16b, v3.16b
        zip1    v2.16b, v2.16b, v4.16b
        addv    h1, v1.8h
        addv    h2, v2.8h
        fmov    w12, s1
        fmov    w13, s2
        orr     w12, w13, w12
        tst     w12, #0xffff
        b.ne    .LBB0_36
        add     x11, x11, #32
        add     x10, x10, #128
        cmp     x11, x9
        b.lo    .LBB0_4
```

> generated code is growing substantially on ARMv7 and to a lesser extent on aarch64

How about setting the CHUNK_SIZE to 9?